Chore: Migrate app-cli command-mknote to TypeScript
The local `note` variable is typed as NoteEntity so the
narrow object-literal shape can be reassigned to the wider
result of Note.save.

Marks the existing fire-and-forget Note.updateGeolocation call
with `void` to satisfy @typescript-eslint/no-floating-promises;
the original JS already ignored the returned promise.

Change by Claude Code
